+++ tests/microservices/Microservices.DicomRelationalMapper.Tests/MicroservicesIntegrationTest.cs
@@ -15,6 +15,7 @@ using Microservices.CohortExtractor.Execution;
 using Microservices.CohortExtractor.Execution.RequestFulfillers;
 using Microservices.DicomRelationalMapper.Execution;
 using Microservices.DicomRelationalMapper.Execution.Namers;
+using Microservices.DicomTagReader;
 using Microservices.DicomTagReader.Execution;
 using Microservices.IdentifierMapper.Execution;
 using Microservices.IdentifierMapper.Execution.Swappers;